returning assets url in graphql. Moving AssetIdentifier logic to indexer-core
