+++ crates/graphql/Cargo.toml
@@ -24,7 +24,6 @@ percent-encoding = "2.1.0"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.70"
 thiserror = "1.0.30"
-
 [dependencies.indexer-core]
 package = "metaplex-indexer-core"
 version = "=0.1.0"
